+++ .make.versions
@@ -85,8 +85,8 @@ CODE2PARQUET_RAY_VERSION=$(DPK_VERSION)
 INGEST_TO_PARQUET_VERSION=$(DPK_VERSION)
 REPO_LVL_ORDER_RAY_VERSION=$(DPK_VERSION)
 
-PDF2MD_PYTHON_VERSION=$(DPK_VERSION)
-PDF2MD_RAY_VERSION=$(DPK_VERSION)
+PDF2PARQUET_PYTHON_VERSION=$(DPK_VERSION)
+PDF2PARQUET_RAY_VERSION=$(DPK_VERSION)
 
 ################## ################## ################## ################## ################## ##################
 # Begin versions that the repo depends on.
