+++ transforms/universal/pdf2parquet/python/src/pdf2parquet_transform_python.py
@@ -1,0 +1,42 @@
+# (C) Copyright IBM Corp. 2024.
+# Licensed under the Apache License, Version 2.0 (the “License”);
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#  http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an “AS IS” BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+################################################################################
+
+import time
+
+from data_processing.runtime.pure_python import PythonTransformLauncher
+from data_processing.runtime.pure_python.runtime_configuration import (
+    PythonTransformRuntimeConfiguration,
+)
+from data_processing.utils import get_logger
+from pdf2parquet_transform import Pdf2ParquetTransformConfiguration
+
+
+logger = get_logger(__name__)
+
+
+class Pdf2ParquetPythonTransformConfiguration(PythonTransformRuntimeConfiguration):
+    """
+    Implements the PythonTransformConfiguration for PDF2PARQUET as required by the PythonTransformLauncher.
+    """
+
+    def __init__(self):
+        """
+        Initialization
+        :param base_configuration - base configuration class
+        """
+        super().__init__(transform_config=Pdf2ParquetTransformConfiguration())
+
+
+if __name__ == "__main__":
+    launcher = PythonTransformLauncher(Pdf2ParquetPythonTransformConfiguration())
+    logger.info("Launching pdf2parquet transform")
+    launcher.launch()
