+++ transforms/universal/pdf2parquet/python/test-data/expected/2206.01062.parquet
