+++ transforms/universal/pdf2parquet/ray/.dockerignore
